requirements.txt
anthropic>=0.39.0
requests>=2.31.0
python-dotenv>=1.0.0

# Vector database and embeddings for RAG
chromadb>=0.4.0
sentence-transformers>=2.2.0

# Document loaders for knowledge base
pypdf>=3.0.0
unstructured
google-generativeai>=0.8.0
slack_bolt>=1.18.0
flask>=3.0.0
mangum>=0.17.0
